requirements.txt
absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
bidict==0.22.1
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cmake==3.25.0
comm==0.1.3
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.5
distro==1.7.0
distro-info==1.1+ubuntu1
dnspython==2.4.2
eventlet==0.33.3
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.9.0
Flask==1.1.2
Flask-Cors==4.0.0
Flask-SocketIO==4.3.1
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
gast==0.5.4
gcloud==0.18.3
google-auth==2.20.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.54.0
gunicorn==21.2.0
gyp==0.1
h5py==3.9.0
httplib2==0.20.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.1.2
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==2.11.3
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jws==0.1.3
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.8.2
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
lit==15.0.7
Markdown==3.4.3
MarkupSafe==1.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.9.0.1
mistune==3.0.1
more-itertools==8.10.0
mpmath==1.2.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.68
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.6.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.18.0
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.42.2
PyJWT==2.4.0
pyparsing==3.0.7
Pyrebase4==4.7.1
pyrsistent==0.19.3
PySocks==1.7.1
python-apt==2.4.0+22.10
python-dateutil==2.8.2
python-engineio==3.13.2
python-json-logger==2.0.7
python-jwt==2.0.1
python-socketio==4.6.0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.11.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
Send2Trash==1.8.2
sentry-sdk==1.21.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
ssh-import-id==5.11
stack-data==0.6.2
sympy==1.11.1
systemd-python==235
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.1
tensorflow-estimator==2.8.0
tensorflow-gpu==2.8.1
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
torchaudio==2.0.1+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.0
tzdata==2023.3
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.16.3+ds
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.6.1
Werkzeug==1.0.1
widgetsnbextension==4.0.7
wrapt==1.15.0
xkit==0.0.0
zipp==3.15.0
